🗑️ refactor(category-filter): consolidate category filter component and remove old implementation
